+++ qcio/models/utils.py
@@ -2,6 +2,7 @@
 
 import importlib
 import warnings
+from collections.abc import Iterable
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
@@ -557,3 +558,15 @@ def rmsd(
         rmsd, _ = rdMolAlign.GetAlignmentTransform(mol2, mol1)
 
     return rmsd
+
+
+def to_multi_xyz(structures: Iterable["Structure"]) -> str:
+    """Create a multi-structure XYZ string from a list of structures.
+
+    Args:
+        structures: An Iterable of Structure objects.
+
+    Returns:
+        The multi-structure XYZ string.
+    """
+    return "".join(struct.to_xyz() for struct in structures)